src/web_app/requirements.txt
flask==3.0.0
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
Werkzeug>=3.0.0
